[FE/#60] Feat :  NewLabel 컴포넌트 분리
