+++ FE/src/components/Labels/Navigation/Navigation.jsx
@@ -4,11 +4,12 @@ import Box from "@material-ui/core/Box";
 
 import CustomButton from "@Components/common/CustomButton";
 import LinkButtons from "@Components/common/LinkButtons/LinkButtons";
-import NewLabel from "./NewLabel";
+import NewLabel from "./NewLabel/NewLabel";
 
 const Navigation = () => {
-  const NEW_LABEL_BTN_TEXT = "New Label";
   const [showNewLabel, setShowNewLabel] = useState(false);
+
+  const NEW_LABEL_BTN_TEXT = "New Label";
 
   return (
     <>
