+++ FE/src/pages/Labels.jsx
@@ -2,7 +2,7 @@ import React, { useContext } from "react";
 
 import Footer from "@Components/common/Footer";
 import LoadingIndicator from "@Components/common/LoadingIndicator";
-import Navigation from "@Components/Labels/Navigation";
+import Navigation from "@Components/Labels/Navigation/Navigation";
 import LabelTable from "@Components/Labels/LabelTable";
 
 import { LabelListContext } from "@Contexts/labelListContext";
